chore: context 관련 코드 분리
