+++ src/components/tabs/Tabs.tsx
@@ -2,6 +2,7 @@ import { Box } from '@components/box/Box';
 import React, { PropsWithChildren } from 'react';
 import { styled } from 'stitches.config';
 import { TabProps } from './Tab';
+import { TabsContext } from './TabsContext';
 
 interface TabsProps {
   text: string;
@@ -9,21 +10,6 @@ interface TabsProps {
   onChange: (text: string) => void;
   children: React.ReactElement<TabProps>[];
 }
-
-interface TabsContextProps {
-  text: string;
-  size: string;
-  onChange: (text: string) => void;
-}
-
-const TabsContext = React.createContext<TabsContextProps>({
-  text: '',
-  size: '',
-  onChange: () => {},
-});
-export const useTabsContext = () => {
-  return React.useContext(TabsContext);
-};
 
 function Tabs({
   text,
